controller: moved isOpenshift function to utils common and export it
