+++ tests/utils.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"net/http"
 	"os"
 	"os/exec"
 	"reflect"
@@ -18,7 +17,6 @@ import (
 
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes"
 
 	"kubevirt.io/containerized-data-importer/tests/framework"
 )
@@ -170,33 +168,4 @@ func PodSpecHasTestNodePlacementValues(f *framework.Framework, podSpec v1.PodSpe
 		return false
 	}
 	return true
-}
-
-//IsOpenshift checks if we are on OpenShift platform
-func IsOpenshift(client kubernetes.Interface) bool {
-	//OpenShift 3.X check
-	result := client.Discovery().RESTClient().Get().AbsPath("/oapi/v1").Do(context.TODO())
-	var statusCode int
-	result.StatusCode(&statusCode)
-
-	if result.Error() == nil {
-		// It is OpenShift
-		if statusCode == http.StatusOK {
-			return true
-		}
-	} else {
-		// Got 404 so this is not Openshift 3.X, let's check OpenShift 4
-		result = client.Discovery().RESTClient().Get().AbsPath("/apis/route.openshift.io").Do(context.TODO())
-		var statusCode int
-		result.StatusCode(&statusCode)
-
-		if result.Error() == nil {
-			// It is OpenShift
-			if statusCode == http.StatusOK {
-				return true
-			}
-		}
-	}
-
-	return false
 }
